Adding django test (#83)
* adding django test

* updating configure.sh

* testing with docker

* adding backend test to jenkins

* Adding dashboard view tests

* fixing mjd test

* adding test coverage

* pep8 and recovering django views

* pep8 scalar_metrics

+++ configure.sh
@@ -18,19 +18,21 @@ cp docker-compose.yml.template docker-compose.yml
 echo 'Copying default frontend .env'
 cp .env.template .env
 
-cd ..
-echo 'Checking spectro test data'
-if [ "$(ls -A backend/spectro/data)" ]; then
-    echo "=> spectro test data check OK"
-else
-    echo 'Downloading spectro test data'
-    cd backend
-    mkdir spectro && cd spectro
-    wget -c http://portal.nersc.gov/project/desi/data/quicklook/20190101_small.tar.gz 
+if [ -z "$1" ]; then
+    cd ..
+    echo 'Checking spectro test data'
+    if [ "$(ls -A backend/spectro/data)" ]; then
+        echo "=> spectro test data check OK"
+    else
+        echo 'Downloading spectro test data'
+        cd backend
+        mkdir spectro && cd spectro
+        wget -c http://portal.nersc.gov/project/desi/data/quicklook/20190101_small.tar.gz 
 
-    echo 'Unzipping...'
-    tar xvzf 20190101_small.tar.gz
-    rm 20190101_small.tar.gz
+        echo 'Unzipping...'
+        tar xvzf 20190101_small.tar.gz
+        rm 20190101_small.tar.gz
+    fi
 fi
 
 echo 'To start run ./start.sh'
